Refactor QR code generation and update RecoveryPhraseQrSheet UI
- Relocate `QRCodeGenerator` to a dedicated `uicomponents.qr` package.
- Extract `QRCodeImage` into a standalone reusable Composable.
- Update `RecoveryPhraseQrSheet` with a new layout, including the mnemonic string display and a debug preview.
- Add temporary debug code in `InitialViewModel` to force the setup state.
